requirements.txt
pdfplumber>=0.11.0
pypdf>=4.0.0
python-docx>=1.0.0
python-multipart>=0.0.9
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
asyncpg>=0.29.0
rank_bm25>=0.2.2
python-dotenv>=1.0.0